NCI-Agency/anet#3320: Fix circular dependency with deserializeQueryParams
