+++ client/src/components/SearchFilters.js
@@ -183,7 +183,7 @@ const advancedSelectFilterTaskProps = {
   addon: TASKS_ICON
 }
 
-const searchFilters = function() {
+export const searchFilters = function() {
   const filters = {}
   const authorWidgetFilters = {
     all: {
@@ -618,6 +618,11 @@ export const SearchDescription = ({ searchQuery, showPlaceholders }) => {
   )
 }
 
+SearchDescription.propTypes = {
+  searchQuery: SearchQueryPropType,
+  showPlaceholders: PropTypes.bool
+}
+
 export const findCommonFiltersForAllObjectTypes = (
   searchObjectTypes,
   theSearchFilters
@@ -628,9 +633,48 @@ export const findCommonFiltersForAllObjectTypes = (
       _pickBy(filters1, (value, key) => filters2[key])
     )
 
-SearchDescription.propTypes = {
-  searchQuery: SearchQueryPropType,
-  showPlaceholders: PropTypes.bool
+export const deserializeQueryParams = (
+  objType,
+  queryParams,
+  callbackFunction
+) => {
+  // From query params to search filters
+  const text = queryParams.text || ""
+  const usedFilters = []
+  const promises = []
+  if (objType) {
+    const EXTRA_FILTERS = extraFilters()
+    const extraFilterDefs = EXTRA_FILTERS[objType] || []
+    extraFilterDefs.map(filterKey => {
+      if (Object.prototype.hasOwnProperty.call(queryParams, filterKey)) {
+        usedFilters.push({ key: filterKey, value: queryParams[filterKey] })
+      }
+      return null
+    })
+    const ALL_FILTERS = searchFilters()
+    const filterDefs = ALL_FILTERS[objType].filters
+    Object.keys(filterDefs).map(filterKey => {
+      const filterDef = filterDefs[filterKey]
+      const deser = filterDef.deserializer(
+        filterDef.props,
+        queryParams,
+        filterKey
+      )
+      if (deser && deser.then instanceof Function) {
+        // deserialize returns a Promise
+        promises.push(deser)
+      } else if (deser) {
+        // deserialize returns filter data
+        usedFilters.push(deser)
+      }
+      return null
+    })
+  }
+  Promise.all(promises).then(dataList => {
+    dataList.forEach((filterData, index) => {
+      // update filters
+      usedFilters.push(filterData)
+    })
+    callbackFunction(objType, usedFilters, text)
+  })
 }
-
-export default { searchFilters, extraFilters }
