+++ client/src/pages/Home.js
@@ -18,7 +18,10 @@ import {
   useBoilerplate
 } from "components/Page"
 import SavedSearchTable from "components/SavedSearchTable"
-import { RECURSE_STRATEGY } from "components/SearchFilters"
+import {
+  deserializeQueryParams,
+  RECURSE_STRATEGY
+} from "components/SearchFilters"
 import { LAST_WEEK } from "dateUtils"
 import _isEmpty from "lodash/isEmpty"
 import { Person, Report } from "models"
@@ -35,7 +38,6 @@ import {
 } from "react-bootstrap"
 import { connect } from "react-redux"
 import { useHistory, useLocation } from "react-router-dom"
-import { deserializeQueryParams } from "searchUtils"
 import Settings from "settings"
 import utils from "utils"
