+++ client/src/pages/reports/Show.js
@@ -16,6 +16,7 @@ import { parseHtmlWithLinkTo } from "components/editor/LinkAnet"
 import LinkTo from "components/LinkTo"
 import Messages from "components/Messages"
 import { DEFAULT_CUSTOM_FIELDS_PARENT } from "components/Model"
+import NoPaginationTaskTable from "components/NoPaginationTaskTable"
 import {
   AnchorLink,
   jumpToTop,
@@ -28,8 +29,8 @@ import RelatedObjectNotes, {
   GRAPHQL_NOTES_FIELDS
 } from "components/RelatedObjectNotes"
 import { ReportFullWorkflow } from "components/ReportWorkflow"
+import { deserializeQueryParams } from "components/SearchFilters"
 import Tag from "components/Tag"
-import NoPaginationTaskTable from "components/NoPaginationTaskTable"
 import { Field, Form, Formik } from "formik"
 import _concat from "lodash/concat"
 import _isEmpty from "lodash/isEmpty"
@@ -44,7 +45,6 @@ import Confirm from "react-confirm-bootstrap"
 import { connect } from "react-redux"
 import { useHistory, useParams } from "react-router-dom"
 import { toast } from "react-toastify"
-import { deserializeQueryParams } from "searchUtils"
 import Settings from "settings"
 import utils from "utils"
 import AuthorizationGroupTable from "./AuthorizationGroupTable"
